исправил небольшие недочеты
