+++ js/main.js
@@ -1,9 +1,1 @@
-import './util.js';
-import './get-photo-data.js';
-
-import './render-big-picture.js';
-import './render-comments.js';
-import {renderThumbnailList} from './render-thumbnails.js';
-renderThumbnailList();
-
-import './form-img-upload';
+import './gallery.js';
